Create a viewModel and a service class
