+++ Movies/Services/Service.swift
@@ -1,0 +1,8 @@
+//
+//  Service.swift
+//  Movies
+//
+//  Created by Victor Hugo Bernardino Magnani on 02/08/21.
+//
+
+import Foundation
