+++ Movies/Utilities/Box.swift
@@ -1,0 +1,8 @@
+//
+//  Box.swift
+//  Movies
+//
+//  Created by Victor Hugo Bernardino Magnani on 02/08/21.
+//
+
+import Foundation
